Add a divider between "Use as" + "Encryption method" and other options
As both these options change what other settings are possible and can
change the defaults of those settings when a different value is
selected, it seems a good idea to add a divider as visual separator.
See discussion in #478598.

For this, crypto_type is moved up from 40 to 20 for the active_partition
dialog. The divider is added at 25.
The second divider (at 75) becomes a simple symlink to the one at 25.

r53533